requirements.txt
# ============================================
# HikVision Bulletproof - Dependencies
# ============================================
# Install with: pip install -r requirements.txt

# HTTP/API Communication
requests>=2.31.0
urllib3>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Database
mysql-connector-python>=8.2.0

# Dashboard
streamlit>=1.30.0
pandas>=2.1.0
plotly>=5.18.0

# Process Management
psutil>=5.9.0

